hook cluster name not found counter elsewhere
